app/requirements.txt
aiohttp==3.8.3
azure-keyvault-secrets==4.6.0
azure-identity==1.12.0
azure-storage-blob==12.14.1
fastapi==0.87.0
uvicorn[standard]==0.20.0